Integrate develop changes (tool version blacklist, GitHub pipeline
selectors, share-with, API tokens, dev proxy, VS Code extension) with
the React 18 / Ant Design 6 upgrade branch.

+++ api/src/test/java/com/epam/pipeline/manager/CmdExecutorTest.java
@@ -43,7 +43,7 @@ public class CmdExecutorTest {
         CmdExecutor cmdExecutor = new CmdExecutor();
 
         // bash sleep measured in sec but timeout in mills
-        double timeToSleep = TIMEOUT / 2 / MILLS_IN_SEC;
+        double timeToSleep = (double) (TIMEOUT / 2) / MILLS_IN_SEC;
         cmdExecutor.executeCommand(String.format(COMMAND_WITH_SLEEP, timeToSleep), TIMEOUT);
 
     }
@@ -53,7 +53,7 @@ public class CmdExecutorTest {
         CmdExecutor cmdExecutor = new CmdExecutor();
 
         // bash sleep measured in sec but timeout in mills
-        double timeToSleep = (TIMEOUT + TIMEOUT) / MILLS_IN_SEC;
+        double timeToSleep = (double) (TIMEOUT + TIMEOUT) / MILLS_IN_SEC;
         cmdExecutor.executeCommand(String.format(COMMAND_WITH_SLEEP, timeToSleep), TIMEOUT);
     }
 }
